+++ src/app/router.tsx
@@ -1,7 +1,10 @@
 import { Route, Routes } from 'react-router-dom'
+import AboutPage from '../pages/AboutPage'
 import LoginPage from '../pages/LoginPage'
 import NotFoundPage from '../pages/NotFoundPage'
+import PrivacyPolicyPage from '../pages/PrivacyPolicyPage'
 import TablePage from '../pages/TablePage'
+import TermsPage from '../pages/TermsPage'
 import Layout from './Layout'
 import ProtectedRoute from './ProtectedRoute'
 import { ROUTES } from './routes'
@@ -19,6 +22,9 @@ function AppRouter() {
             </ProtectedRoute>
           }
         />
+        <Route path={ROUTES.about} element={<AboutPage />} />
+        <Route path={ROUTES.privacy} element={<PrivacyPolicyPage />} />
+        <Route path={ROUTES.terms} element={<TermsPage />} />
         <Route path="*" element={<NotFoundPage />} />
       </Route>
     </Routes>